Updated the search to work with a places array.
HealthMap alerts can have different values for different geographic locations.  If this data is stored in the meta.places list, this will now be included in a search.  Also, there can be multiple diseases within the alert.  If this list is stored in meta.diseases or in meta.places..diseases, all of the diseases will be searched.

Internally, when searching for a country, disease, or species, the match can take place on any value within the item.  If searching for two or more of these, an item will match even if it has these values from different places.  For example, in alert 2932898 (internal name 29328980000) there is an entry for country: United States, disease: Enterovirus 68, and another entry for country: Saudi Arabia, disease: MERS.  Seaching for country: United States, disease: MERS will return this entry.