Added an mongo index to speed up searches.
If you tried to return more than a thousand or so records, mongo with report an error.  This is fixed by having an index on meta.date.  We could do more with indices to speed up searches.

Added a plugin version number.

Changed ModelImporter() to ModelImporter; there is no need to instantiate the class to access a static method.